+++ internal/errors/errors.go
@@ -3,6 +3,8 @@ package errors
 import (
 	"errors"
 	"fmt"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 // -----------------------------------------------------------------------------
@@ -23,12 +25,31 @@ type ErrUnsupportedGatewayClass struct {
 	reason string
 }
 
+// NewErrUnsupportedGateway creates a new ErrUnsupportedGatewayClass error
 func NewErrUnsupportedGateway(reason string) ErrUnsupportedGatewayClass {
 	return ErrUnsupportedGatewayClass{reason: reason}
 }
 
+// Error returns the error message for the ErrUnsupportedGatewayClass error
 func (e ErrUnsupportedGatewayClass) Error() string {
 	return fmt.Sprintf("unsupported gateway class: %s", e.reason)
+}
+
+// ErrNotAcceptedGatewayClass is an error which indicates that a provided GatewayClass
+// is not accepted.
+type ErrNotAcceptedGatewayClass struct {
+	gatewayClass string
+	condition    metav1.Condition
+}
+
+// NewErrNotAcceptedGatewayClass creates a new ErrNotAcceptedGatewayClass error
+func NewErrNotAcceptedGatewayClass(gatewayClass string, condition metav1.Condition) ErrNotAcceptedGatewayClass {
+	return ErrNotAcceptedGatewayClass{gatewayClass: gatewayClass, condition: condition}
+}
+
+// Error returns the error message for the ErrNotAcceptedGatewayClass error
+func (e ErrNotAcceptedGatewayClass) Error() string {
+	return fmt.Sprintf("gateway class %s not accepted; reason: %s, message: %s", e.gatewayClass, e.condition.Reason, e.condition.Message)
 }
 
 // -----------------------------------------------------------------------------